+++ NEWS.md
@@ -22,6 +22,8 @@
 
   `mkvpropedit` and the GUI's header editor are unaffected as they've always
   allowed to set the flag on as many tracks as the user wanted.
+* mkvinfo: the option `-X`/`--full-hexdump` now affects all binary elements,
+  not just the frame data in `SimpleBlock` and `BlockGroup` elements.
 
 
 ## Bug fixes